moved event system and global npc to events directory
